Address move to ulong

+++ LogBins/Base/BagAddress.cs
@@ -7,6 +7,6 @@ namespace LogBins.Base
     public struct BagAddress
     {
         public ushort TrainId { get; set; }
-        public int BagId { get; set; }
+        public ushort BagId { get; set; }
     }
 }
